Activate continuous development and remote builds in Okteto Cloud by default
